(feature): folder structure reset
