+++ server/dummy.js
@@ -1,4 +1,4 @@
-const db = require('./db');
+const db = require('./config/db');
 
 db.serialize(() => {
     db.run('DELETE FROM users');
