+++ server/routes/user.routes.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const router = express.Router();
+const userController = require('../controllers/user.controller');
+
+router.post('/add', userController.addUser);
+router.get('/statistics', userController.getStatistics);
+
+module.exports = router;
